refactor: consolidate `bindings.rs` into `abi.rs`
The `bindings.rs` and `abi.rs` files both defined `sol! { contract ZonePortal { ... } }`,
with the `abi.rs` version being a strict superset. This removes the duplication by moving
the `enabled_tokens()` helper and missing events (`EncryptedDepositMade`, `TokenEnabled`)
into `abi.rs`, updating all consumers, and deleting `bindings.rs`.
